Get/Delete ById expected Long but got Integer fixed

+++ src/main/java/autopartner/controller/rest/CarBrandController.java
@@ -31,7 +31,7 @@ public class CarBrandController {
 
     @Secured({"ROLE_ADMIN", "ROLE_ROOT"})
     @RequestMapping(value = "/api/car/brand/{id}", method = RequestMethod.GET)
-    public ResponseEntity<?> get(@PathVariable Integer id) {
+    public ResponseEntity<?> get(@PathVariable Long id) {
         return ResponseEntity.ok(carBrandService.getCarBrandById(id));
     }
