+++ src/main/java/autopartner/repository/CarTypeRepository.java
@@ -3,6 +3,6 @@ package autopartner.repository;
 import autopartner.domain.entity.CarType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarTypeRepository extends CrudRepository<CarType, Integer> {
+public interface CarTypeRepository extends CrudRepository<CarType, Long> {
     Iterable<CarType> findByActiveTrue();
 }
