+++ src/main/java/autopartner/repository/OrderTaskRepository.java
@@ -3,7 +3,7 @@ package autopartner.repository;
 import autopartner.domain.entity.OrderTask;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderTaskRepository extends CrudRepository<OrderTask, Integer> {
+public interface OrderTaskRepository extends CrudRepository<OrderTask, Long> {
     Iterable<OrderTask> findByActiveTrue();
 
     Iterable<OrderTask> findByOrderAndActiveTrue(Long orderId);
