+++ src/main/java/autopartner/repository/TaskTypeRepository.java
@@ -3,6 +3,6 @@ package autopartner.repository;
 import autopartner.domain.entity.TaskType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TaskTypeRepository extends CrudRepository<TaskType, Integer> {
+public interface TaskTypeRepository extends CrudRepository<TaskType, Long> {
     Iterable<TaskType> findByActiveTrue();
 }
