+++ src/main/java/autopartner/service/TaskTypeService.java
@@ -6,7 +6,7 @@ public interface TaskTypeService {
 
     Iterable<TaskType> getByActiveTrue();
 
-    TaskType getTaskTypeById(Integer id);
+    TaskType getTaskTypeById(Long id);
 
     TaskType saveTaskType(TaskType taskType);
